Move the test only `wrap` utilities out of the main bundle
