+++ src/__tests__/ApolloClient.ts
@@ -6,7 +6,7 @@ import { HttpLink } from '../link/http/HttpLink';
 import { InMemoryCache } from '../cache/inmemory/inMemoryCache';
 import { makeReference } from '../cache/inmemory/helpers';
 import { stripSymbols } from './utils/stripSymbols';
-import { withWarning } from '../util/wrap';
+import { withWarning } from './utils/wrap';
 import { ApolloClient } from '../';
 import { DefaultOptions } from '../ApolloClient';
 import { FetchPolicy, QueryOptions } from '../core/watchQueryOptions';
