+++ src/core/__tests__/ObservableQuery.ts
@@ -14,7 +14,7 @@ import { QueryManager } from '../QueryManager';
 import { DataStore } from '../../data/store';
 import { ApolloClient } from '../../';
 
-import wrap from '../../util/wrap';
+import wrap from '../../__tests__/utils/wrap';
 import subscribeAndCount from '../../__tests__/utils/subscribeAndCount';
 import { stripSymbols } from '../../__tests__/utils/stripSymbols';
 import { ApolloError } from '../../errors/ApolloError';
